Refactor common fields into a base data model
