+++ module/documents/items/embedded-feature-data-model.mjs
@@ -1,0 +1,26 @@
+/**
+ * @description Embeds a feature inside a data property
+ */
+export class EmbeddedFeatureDataModel extends foundry.abstract.TypeDataModel {
+	static defineSchema() {
+		const { StringField, SchemaField, BooleanField } = foundry.data.fields;
+		return {
+			fuid: new StringField(),
+			summary: new SchemaField({ value: new StringField() }),
+			source: new StringField(),
+			isFavored: new SchemaField({ value: new BooleanField() }),
+		};
+	}
+
+	prepareDerivedData() {
+		this.data?.prepareData();
+	}
+
+	/**
+	 * For default item chat messages to pick up description.
+	 * @return {*}
+	 */
+	get description() {
+		return this.data.description;
+	}
+}
